move logic out of main function
The majority of the logic has been moved from main() into
filter_map_events* functions. These functions allow a user to pass a
closure, which operates on events.

Data structure for buffering events changed from VecDeque to BTreeMap,
as it now has the additional responsibility of mantaining the (time)
sorted order of events.